atualização na rotas, a criação de produtos foi adicionada a rotas produtos
