+++ routes/login.js
@@ -13,9 +13,10 @@ router.post('/', function(req, res) {
     let user = model.authenticator(email, password)
     if (user === undefined) {
         res.redirect('/login');
+        return;
     }
     req.session.user = user;
-    res.redirect('/novo-produto');
+    res.redirect('/products/create');
 })
 
 module.exports = router;
