+++ views/index.ejs
@@ -16,8 +16,9 @@
             <li><a href="/about">Sobre</a></li>
             <li><a href="/products">Produtos</a> 
             <ul>
+                <li><a href="/products/create">Cadastrar novo produto</a></li>
                 <li><a href="/login">Login In</a></li>
-                <li><a href="/register">Cadastro</a></li>
+                <li><a href="/register">Criar conta</a></li>
             </ul>
             </li></ul>
         </nav>
