+++ views/newProduct.ejs
@@ -19,8 +19,9 @@
                 <li><a href="/about">Sobre</a></li>
                 <li><a href="/products">Você está aqui</a> 
                 <ul>
+                    <li><a href="/products/create">Cadastrar novo produto</a></li>
                     <li><a href="/login">Login In</a></li>
-                    <li><a href="/register">Cadastro</a></li>
+                    <li><a href="/register">Criar conta</a></li>
                 </ul>
                 </li></ul>
             </nav>
@@ -71,11 +72,10 @@
             <div class="container-product">
             <div class="col-12 mt-4"><h3>Lista de Produtos</h3></div>
             
-                    <% const productsList = JSON.parse(products); %>
-                    <% productsList.forEach(function(item) {%>
+                    <% products.forEach(function(item) {%>
                         <div class="col-lg-4 item-product mt-4">    
                         <p> <%= item.productName %> </p>
-                        <img <%= "src=images/uploads/" + item.photo %> >
+                        <img <%= "src=../images/uploads/" + item.photo %> >
                         </div>
                     <%}) %>
